+++ Z/Data/CBytes.hs
@@ -5,6 +5,7 @@
 {-# LANGUAGE UnliftedFFITypes #-}
 {-# LANGUAGE DeriveDataTypeable #-}
 {-# LANGUAGE BangPatterns #-}
+{-# LANGUAGE PatternSynonyms #-}
 {-|
 Module      : Z.Data.CBytes
 Description : Null-ternimated byte string.
@@ -20,16 +21,15 @@ Portability : non-portable
 -}
 
 module Z.Data.CBytes
-  ( CBytes(..)
+  ( CBytes(CB)
+  , toPrimArray
   , pack
   , unpack
   , null , length
   , empty, append, concat, intercalate, intercalateElem
   , toBytes, fromBytes, toText, toTextMaybe, fromText
-  , fromCString, fromCString', fromCStringN
+  , fromCString, fromCStringN
   , withCBytesUnsafe, withCBytes, allocCBytesUnsafe, allocCBytes
-  -- * exception
-  , NullPointerException(..)
   -- re-export
   , CString
   , V.c2w, V.w2c
@@ -64,11 +64,10 @@ import           Prelude                 hiding (all, any, appendFile, break,
                                           unlines, unzip, writeFile, zip,
                                           zipWith)
 import           Z.Data.Array
-import           Z.Data.Array.UnalignedAccess
+import           Z.Data.Array.Unaligned
 import qualified Z.Data.Text           as T
 import           Z.Data.Text.UTF8Codec (encodeCharModifiedUTF8, decodeChar)
 import qualified Z.Data.Vector.Base    as V
-import qualified Z.Data.Vector.Extra   as V
 import           Z.Foreign
 import           System.IO.Unsafe        (unsafeDupablePerformIO)
 
@@ -81,8 +80,8 @@ import           System.IO.Unsafe        (unsafeDupablePerformIO)
 -- interface, e.g. libuv do this when deal with file paths.
 --
 -- 'CBytes' don't support O(1) slicing, it's not suitable to use it to store large byte
--- chunk, we do save the length info on GHC heap though. If you need advance editing,
--- convert 'CBytes' to \/ from 'V.Bytes' with 'toBytes\/fromBytes' and use vector combinators.
+-- chunk, If you need advance editing, convert 'CBytes' to \/ from 'PrimArray' with 'CB',
+-- or 'V.Bytes' with 'toBytes\/fromBytes' if you need O(1) slicing, then use vector combinators.
 --
 -- When textual represatation is needed(conver to 'String', 'T.Text', 'Show' instance, etc.),
 -- we assume 'CBytes' using UTF-8 encodings, 'CBytes' can be used with @OverloadedString@,
@@ -95,14 +94,25 @@ import           System.IO.Unsafe        (unsafeDupablePerformIO)
 -- various places.
 -- --
 data CBytes = CBytes
-    { rawCBytes :: {-# UNPACK #-} !(PrimArray Word8)    -- ^ On heap 'PrimArray'
+    { toPrimArray :: {-# UNPACK #-} !(PrimArray Word8)    -- ^ Convert to 'PrimArray'
     }                                                   -- there's an invariance that this array's
                                                         -- length is always shrinked to contain content
-                                                        -- and \NUL terminator
+                                                        -- before \NUL terminator
 
--- | All exception can be throw by using 'CBytes'.
-data CBytesException = NULLTerminatorNotFound CallStack deriving Show
-instance Exception CBytesException
+pattern CB :: PrimArray Word8 -> CBytes
+pattern CB arr <- CBytes arr where
+    CB arr = fromPrimArray arr
+
+fromPrimArray :: PrimArray Word8 -> CBytes
+{-# INLINE fromPrimArray #-}
+fromPrimArray arr = runST (
+    case V.elemIndex 0 arr of
+        Just i -> do
+            mpa <- newPrimArray i
+            copyPrimArray mpa 0 arr 0 i
+            pa <- unsafeFreezePrimArray mpa
+            return (CBytes pa)
+        _ -> return (CBytes arr))
 
 instance Show CBytes where
     showsPrec p t = showsPrec p (unpack t)
@@ -135,30 +145,37 @@ instance Monoid CBytes where
 
 instance Hashable CBytes where
     hashWithSalt salt (CBytes pa@(PrimArray ba#)) = unsafeDupablePerformIO $ do
-        V.c_fnv_hash_ba ba# 0 (sizeofPrimArray pa - 1) salt
+        V.c_fnv_hash_ba ba# 0 (sizeofPrimArray pa) salt
 
-instance UnalignedAccess CBytes where
+-- | This instance peek bytes until \NUL, poke bytes with an extra \NUL terminator.
+instance Unaligned CBytes where
     {-# INLINE unalignedSize #-}
-    unalignedSize (CBytes arr ) = sizeofPrimArray arr
+    unalignedSize (CBytes arr) = sizeofPrimArray arr + 1
     {-# INLINE peekMBA #-}
     peekMBA mba# i = do
-        l <- hs_strlen_mba mba# i
-        let l' = l+1
-        mpa <- newPrimArray l'
-        copyMutablePrimArray mpa 0 (MutablePrimArray mba#) i l'
+        b <- getSizeofMutableByteArray (MutableByteArray mba#)
+        l <- c_memchr mba# i 0 b
+        let l' = if l == -1 then b else l
+        mpa <- newPrimArray l
+        copyMutablePrimArray mpa 0 (MutablePrimArray mba#) i l
         pa <- unsafeFreezePrimArray mpa
         return (CBytes pa)
+
     {-# INLINE pokeMBA #-}
     pokeMBA mba# i (CBytes pa) = do
-        copyPrimArray (MutablePrimArray mba#) i pa 0 (sizeofPrimArray pa)
+        let l = sizeofPrimArray pa
+        copyPrimArray (MutablePrimArray mba# :: MutablePrimArray RealWorld Word8) i pa 0 l
+        writePrimArray (MutablePrimArray mba# :: MutablePrimArray RealWorld Word8) (i+l) 0
+
     {-# INLINE indexBA #-}
-    indexBA ba# i = unsafeDupablePerformIO $ do
-        l <- hs_strlen_ba ba# i
-        let l' = l+1
-        mpa <- newPrimArray l'
-        copyPrimArray mpa 0 (PrimArray ba#) i l'
+    indexBA ba# i = runST (do
+        let b = sizeofByteArray (ByteArray ba#)
+            l = V.c_memchr ba# i 0 b
+            l' = if l == -1 then b else l
+        mpa <- newPrimArray l
+        copyPrimArray mpa 0 (PrimArray ba#) i l
         pa <- unsafeFreezePrimArray mpa
-        return (CBytes pa)
+        return (CBytes pa))
 
 append :: CBytes -> CBytes -> CBytes
 {-# INLINABLE append #-}
@@ -166,10 +183,9 @@ append strA@(CBytes pa) strB@(CBytes pb)
     | lenA == 0 = strB
     | lenB == 0 = strA
     | otherwise = unsafeDupablePerformIO $ do
-        mpa <- newPrimArray (lenA+lenB+1)
+        mpa <- newPrimArray (lenA+lenB)
         copyPrimArray mpa 0    pa 0 lenA
         copyPrimArray mpa lenA pb 0 lenB
-        writePrimArray mpa (lenA + lenB) 0     -- the \NUL terminator
         pa' <- unsafeFreezePrimArray mpa
         return (CBytes pa')
   where
@@ -177,11 +193,9 @@ append strA@(CBytes pa) strB@(CBytes pb)
     lenB = length strB
 
 -- | 'empty' 'CBytes'
---
--- Passing 'empty' to C FFI is equivalent to passing a @NULL@ pointer.
 empty :: CBytes
 {-# NOINLINE empty #-}
-empty = CBytes (V.singleton 0)
+empty = CBytes (V.empty)
 
 concat :: [CBytes] -> CBytes
 {-# INLINABLE concat #-}
@@ -189,9 +203,8 @@ concat bss = case pre 0 0 bss of
     (0, _) -> empty
     (1, _) -> let Just b = List.find (not . null) bss in b -- there must be a not empty CBytes
     (_, l) -> runST $ do
-        buf <- newPrimArray (l+1)
+        buf <- newPrimArray l
         copy bss 0 buf
-        writePrimArray buf l 0 -- the \NUL terminator
         CBytes <$> unsafeFreezePrimArray buf
   where
     -- pre scan to decide if we really need to copy and calculate total length
@@ -222,16 +235,16 @@ intercalate s = concat . List.intersperse s
 
 -- | /O(n)/ An efficient way to join 'CByte' s with a byte.
 --
--- @\NUL@ will be encoded as @C0 80@.
+-- Intercalate bytes list with @\NUL@ will effectively leave the first bytes.
 intercalateElem :: Word8 -> [CBytes] -> CBytes
 {-# INLINABLE intercalateElem #-}
-intercalateElem 0 bss = intercalate (CBytes (V.pack [0xC0, 0x80, 0x00])) bss
+intercalateElem 0 [] = empty
+intercalateElem 0 (bs:_) = bs
 intercalateElem w8 bss = case len bss 0 of
     0 -> empty
     l -> runST $ do
-        buf <- newPrimArray (l+1)
+        buf <- newPrimArray l
         copy bss 0 buf
-        writePrimArray buf l 0 -- the \NUL terminator
         CBytes <$> unsafeFreezePrimArray buf
   where
     len []     !acc = acc
@@ -265,14 +278,14 @@ instance IsString CBytes where
 packAddr :: Addr# -> CBytes
 packAddr addr0# = go addr0#
   where
-    len = 1 + (fromIntegral . unsafeDupablePerformIO $ V.c_strlen addr0#)
+    len = (fromIntegral . unsafeDupablePerformIO $ V.c_strlen addr0#)
     go addr# = runST $ do
         marr <- newPrimArray len
         copyPtrToMutablePrimArray marr 0 (Ptr addr#) len
         arr <- unsafeFreezePrimArray marr
         return (CBytes arr)
 
--- | Pack a 'String' into null-terminated 'CBytes'.
+-- | Pack a 'String' into 'CBytes'.
 --
 -- @\NUL@ is encoded as two bytes @C0 80@ , '\xD800' ~ '\xDFFF' is encoded as a three bytes normal UTF-8 codepoint.
 pack :: String -> CBytes
@@ -280,8 +293,7 @@ pack :: String -> CBytes
 pack s = runST $ do
     mba <- newPrimArray V.defaultInitSize
     (SP2 i mba') <- foldlM go (SP2 0 mba) s
-    writePrimArray mba' i 0     -- the \NUL terminator
-    shrinkMutablePrimArray mba' (i+1)
+    shrinkMutablePrimArray mba' i
     ba <- unsafeFreezePrimArray mba'
     return (CBytes ba)
   where
@@ -290,7 +302,7 @@ pack s = runST $ do
     go :: SP2 s -> Char -> ST s (SP2 s)
     go (SP2 i mba) !c     = do
         siz <- getSizeofMutablePrimArray mba
-        if i < siz - 4  -- we need at least 5 bytes for safety due to extra '\0' byte
+        if i < siz - 3  -- we need at least 4 bytes for safety
         then do
             i' <- encodeCharModifiedUTF8 mba i c
             return (SP2 i' mba)
@@ -303,7 +315,10 @@ pack s = runST $ do
 
 data SP2 s = SP2 {-# UNPACK #-}!Int {-# UNPACK #-}!(MutablePrimArray s Word8)
 
--- | /O(n)/ Convert cbytes to a char list.
+-- | /O(n)/ Convert cbytes to a char list using UTF8 encoding assumption.
+--
+-- This function is much tolerant than 'toText', it simply decoding codepoints using UTF8 'decodeChar'
+-- without checking, thus @pack . unpack /== id@ ('pack' will replace illegal codepoints with replacement char).
 --
 -- Unpacking is done lazily. i.e. we will retain reference to the array until all element are consumed.
 --
@@ -312,7 +327,7 @@ unpack :: CBytes -> String
 {-# INLINE [1] unpack #-}
 unpack (CBytes arr) = go 0
   where
-    !end = sizeofPrimArray arr - 1
+    !end = sizeofPrimArray arr
     go !idx
         | idx >= end = []
         | otherwise = let (# c, i #) = decodeChar arr idx in c : go (idx + i)
@@ -321,7 +336,7 @@ unpackFB :: CBytes -> (Char -> a -> a) -> a -> a
 {-# INLINE [0] unpackFB #-}
 unpackFB (CBytes arr) k z = go 0
   where
-    !end = sizeofPrimArray arr - 1
+    !end = sizeofPrimArray arr
     go !idx
         | idx >= end = z
         | otherwise = let (# c, i #) = decodeChar arr idx in c `k` go (idx + i)
@@ -337,39 +352,37 @@ unpackFB (CBytes arr) k z = go 0
 --
 null :: CBytes -> Bool
 {-# INLINE null #-}
-null (CBytes pa) = indexPrimArray pa 0 == 0
+null (CBytes pa) = sizeofPrimArray pa == 0
 
 -- Return the BTYE length of 'CBytes'.
 --
 length :: CBytes -> Int
 {-# INLINE length #-}
-length (CBytes pa) = sizeofPrimArray pa - 1
+length (CBytes pa) = sizeofPrimArray pa
 
 -- | /O(1)/, convert to 'V.Bytes', which can be processed by vector combinators.
---
--- NOTE: the @\NUL@ ternimator is not included.
 toBytes :: CBytes -> V.Bytes
 {-# INLINABLE toBytes #-}
-toBytes (CBytes arr) = V.PrimVector arr 0 (sizeofPrimArray arr - 1)
+toBytes (CBytes arr) = V.PrimVector arr 0 (sizeofPrimArray arr)
 
 -- | /O(n)/, convert from 'V.Bytes'
 --
--- Result will be trimmed down to first @\NUL@ byte if there's any, or
--- appends a @\NUL@ byte if there's none.
+-- Result will be trimmed down to first byte before @\NUL@ byte if there's any.
 fromBytes :: V.Bytes -> CBytes
 {-# INLINABLE fromBytes #-}
 fromBytes v@(V.PrimVector arr s l) = runST (do
-    if s == 0 && sizeofPrimArray arr == l+1 && indexPrimArray arr l == 0 -- there's \NUL already
-    then return (CBytes arr)
-    else do
-        let l' = case V.elemIndex 0 v of
-                    Just i -> i
-                    _ -> l
-        mpa <- newPrimArray (l'+1)
-        copyPrimArray mpa 0 arr s l'
-        writePrimArray mpa l' 0     -- the \NUL terminator
-        pa <- unsafeFreezePrimArray mpa
-        return (CBytes pa))
+    case V.elemIndex 0 v of
+        Just i -> do
+            mpa <- newPrimArray i
+            copyPrimArray mpa 0 arr s i
+            pa <- unsafeFreezePrimArray mpa
+            return (CBytes pa)
+        _ | s == 0 && sizeofPrimArray arr == l -> return (CBytes arr)
+          | otherwise -> do
+                mpa <- newPrimArray l
+                copyPrimArray mpa 0 arr s l
+                pa <- unsafeFreezePrimArray mpa
+                return (CBytes pa))
 
 -- | /O(n)/, convert to 'T.Text' using UTF8 encoding assumption.
 --
@@ -387,8 +400,7 @@ toTextMaybe = T.validateMaybe . toBytes
 
 -- | /O(n)/, convert from 'T.Text',
 --
--- Result will be trimmed down to first @\NUL@ byte if there's any, or
--- appends a @\NUL@ byte if there's none.
+-- Result will be trimmed down to first byte before @\NUL@ byte if there's any.
 fromText :: T.Text -> CBytes
 {-# INLINABLE fromText #-}
 fromText = fromBytes . T.getUTF8Bytes
@@ -405,32 +417,14 @@ fromCString cstring = do
     then return empty
     else do
         len <- fromIntegral <$> c_strlen_ptr cstring
-        let len' = len + 1
-        mpa <- newPrimArray len'
-        copyPtrToMutablePrimArray mpa 0 (castPtr cstring) len'
+        mpa <- newPrimArray len
+        copyPtrToMutablePrimArray mpa 0 (castPtr cstring) len
         pa <- unsafeFreezePrimArray mpa
         return (CBytes pa)
 
--- | Same with 'fromCString', but throw 'NullPointerException' when meet a null pointer.
---
-fromCString' :: HasCallStack => CString -> IO (Maybe CBytes)
-{-# INLINABLE fromCString' #-}
-fromCString' cstring =
-    if cstring == nullPtr
-    then throwIO (NullPointerException callStack)
-    else do
-        len <- fromIntegral <$> c_strlen_ptr cstring
-        let len' = len + 1
-        mpa <- newPrimArray len'
-        copyPtrToMutablePrimArray mpa 0 (castPtr cstring) len'
-        pa <- unsafeFreezePrimArray mpa
-        return (Just (CBytes pa))
-
 -- | Same with 'fromCString', but only take at most N bytes.
 --
--- Result will be trimmed down to first @\NUL@ byte if there's any, or
--- appends a @\NUL@ byte if there's none.
---
+-- Result will be trimmed down to first byte before @\NUL@ byte if there's any.
 fromCStringN :: CString -> Int -> IO CBytes
 {-# INLINABLE fromCStringN #-}
 fromCStringN cstring len0 = do
@@ -439,70 +433,73 @@ fromCStringN cstring len0 = do
     else do
         len1 <- fromIntegral <$> c_strlen_ptr cstring
         let len = min len0 len1
-        mpa <- newPrimArray (len+1)
+        mpa <- newPrimArray len
         copyPtrToMutablePrimArray mpa 0 (castPtr cstring) len
-        writePrimArray mpa len 0     -- the \NUL terminator
         pa <- unsafeFreezePrimArray mpa
         return (CBytes pa)
-
-data NullPointerException = NullPointerException CallStack deriving (Show, Typeable)
-instance Exception NullPointerException
 
 -- | Pass 'CBytes' to foreign function as a @const char*@.
 --
 -- USE THIS FUNCTION WITH UNSAFE FFI CALL ONLY.
 withCBytesUnsafe :: CBytes -> (BA# Word8 -> IO a) -> IO a
 {-# INLINABLE withCBytesUnsafe #-}
-withCBytesUnsafe (CBytes pa) f = withPrimArrayUnsafe pa (\ p _ -> f p)
+withCBytesUnsafe (CBytes pa) f = do
+    let l = sizeofPrimArray pa
+    mpa <- newPrimArray (l+1)
+    copyPrimArray mpa 0 pa 0 l
+    writePrimArray mpa l 0
+    pa' <- unsafeFreezePrimArray mpa
+    withPrimArrayUnsafe pa' (\ p _ -> f p)
 
 -- | Pass 'CBytes' to foreign function as a @const char*@.
 --
 -- Don't pass a forever loop to this function, see <https://ghc.haskell.org/trac/ghc/ticket/14346 #14346>.
 withCBytes :: CBytes -> (Ptr Word8 -> IO a) -> IO a
 {-# INLINABLE withCBytes #-}
-withCBytes (CBytes pa) f = withPrimArraySafe pa (\ p _ -> f p)
+withCBytes (CBytes pa) f = do
+    let l = sizeofPrimArray pa
+    mpa <- newPinnedPrimArray (l+1)
+    copyPrimArray mpa 0 pa 0 l
+    writePrimArray mpa l 0
+    pa' <- unsafeFreezePrimArray mpa
+    withPrimArraySafe pa' (\ p _ -> f p)
 
 -- | Create a 'CBytes' with IO action.
 --
 -- If (<=0) capacity is provided, a pointer pointing to @\NUL@ is passed to initialize function
 -- and 'empty' will be returned. Which may cause troubles for some FFI functions.
--- Other than that, User have to make sure a @\NUL@ ternimated
--- string will be written, otherwise a 'NULLTerminatorNotFound' will be thrown.
 allocCBytesUnsafe :: HasCallStack
-                  => Int  -- ^ capacity n, including the @\NUL@ terminator
+                  => Int                   -- ^ capacity n(not include the @\NUL@ terminator)
                   -> (MBA# Word8 -> IO a)  -- ^ initialization function,
                   -> IO (CBytes, a)
 {-# INLINABLE allocCBytesUnsafe #-}
 allocCBytesUnsafe n fill | n <= 0 = withPrimUnsafe (0::Word8) fill >>=
                                         \ (_, b) -> return (empty, b)
                          | otherwise = do
-    mba@(MutablePrimArray mba#) <- newPrimArray n :: IO (MutablePrimArray RealWorld Word8)
+    mba@(MutablePrimArray mba#) <- newPrimArray (n+1) :: IO (MutablePrimArray RealWorld Word8)
     a <- fill mba#
-    l <- fromIntegral <$> (c_strlen_mba mba#)
-    when (l+1>n) (throwIO (NULLTerminatorNotFound callStack))
-    shrinkMutablePrimArray mba (l+1)
+    l <- fromIntegral <$> (c_memchr mba# 0 0 n)
+    shrinkMutablePrimArray mba (if l == -1 then n else l)
     bs <- unsafeFreezePrimArray mba
     return (CBytes bs, a)
-
 
 
 -- | Create a 'CBytes' with IO action.
 --
 -- If (<=0) capacity is provided, a 'nullPtr' is passed to initialize function and
 -- 'empty' will be returned. Other than that, User have to make sure a @\NUL@ ternimated
--- string will be written, otherwise a 'NULLTerminatorNotFound' will be thrown.
+-- string will be written.
 allocCBytes :: HasCallStack
-            => Int  -- ^ capacity n, including the @\NUL@ terminator
+            => Int                -- ^ capacity n(not include the @\NUL@ terminator)
             -> (CString -> IO a)  -- ^ initialization function,
             -> IO (CBytes, a)
 {-# INLINABLE allocCBytes #-}
 allocCBytes n fill | n <= 0 = fill nullPtr >>= \ a -> return (empty, a)
                    | otherwise = do
-    mba <- newPinnedPrimArray n :: IO (MutablePrimArray RealWorld Word8)
+    mba@(MutablePrimArray mba#) <- newPinnedPrimArray (n+1) :: IO (MutablePrimArray RealWorld Word8)
     a <- withMutablePrimArrayContents mba (fill . castPtr)
-    l <- fromIntegral <$> withMutablePrimArrayContents mba (c_strlen_ptr . castPtr)
-    when (l+1>n) (throwIO (NULLTerminatorNotFound callStack))
-    shrinkMutablePrimArray mba (l+1)
+    l <- fromIntegral <$> (c_memchr mba# 0 0 n)
+    shrinkMutablePrimArray mba (if l == -1 then n else l)
     bs <- unsafeFreezePrimArray mba
     return (CBytes bs, a)
 
@@ -512,6 +509,5 @@ c_strlen_ptr :: CString -> IO CSize
 {-# INLINE c_strlen_ptr #-}
 c_strlen_ptr (Ptr a#) = V.c_strlen a#
 
-foreign import ccall unsafe "string.h strlen" c_strlen_mba :: MBA# Word8 -> IO CSize
-foreign import ccall unsafe "string.h hs_strlen" hs_strlen_mba :: MBA# Word8 -> Int -> IO Int
-foreign import ccall unsafe "string.h hs_strlen" hs_strlen_ba :: BA# Word8 -> Int -> IO Int
+-- HsInt hs_memchr(uint8_t *a, HsInt aoff, uint8_t b, HsInt n);
+foreign import ccall unsafe "hs_memchr" c_memchr :: MBA# Word8 -> Int -> Word8 -> Int -> IO Int